backend/app/requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
sqlalchemy==2.0.35
python-jose==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
email-validator==2.2.0
pip install python-multipart